requirements.txt
Cython
EasyDict
opencv-python
mxnet-cu101
mxboard
pyyaml